Eliminate need for extra package restore step
Per the doc updates to NuProj found here:
https://github.com/nuproj/nuproj/commit/883121f807232dc6df0c0b6e5353f6c0215d99ca
We don't need to run "nuget restore project.json" on the NuProj directory directly any more. Using a solution-level packages.config file solves it in a simpler way.

Yes, this does mean that each solution has a local copy of the NuProj package after restore completes, but that's not too bad. And at least for this repo, that means that "nuget restore" on the solution always just works. And that's valuable.