Enable JSON-RPC servers to completely control their error response
Servers could already control the error.message property via the Message property on the exception they throw.
With this change, they can also control the error.code and error.data properties, by throwing a LocalRpcException which defines properties for these two other fields.

This also replaces the RemoteStackTrace and RemoteErrorCode properties which assumed a particular schema for the error.data object in favor of a new type that servers and clients can *optionally* use to convey this and more information.
The default behavior continues to include the information on the server we had before, plus the exception *type*.

Fixes #65