More confidently fault aborted requests
The problem was we started shutting things down before faulting requests, such that a race existed in which requests might be "canceled" before we got around to faulting them. So with this change, we fault them *first*, then shut other stuff down.

This should stabilize the JsonRpcWithFatalExceptionsTests tests that are sporatically failing ever since 25a8e35ee was checked in.