Avoid returning object from "async Task" (non-generic) server methods
In order to be resilient to changes across .NET Framework, .NET Core, and various versions thereof, rather than special case detection of `Task<VoidTaskResult>`, we simply follow the same pattern that C# uses: if you want to return a value, return `Task<T>` instead of a non-generic `Task`.

This is technically a breaking change, however, as is demonstrated by two tests I had to fix up. For example, a `Task` returning method that actually returned an instance of `Task<T>` (which derives from `Task`) will no longer return the task's result to the client.

Fixes #259