Avoid linking CancellationTokens in MessageHandlerBase
It might be nice to forcefully cancel folks when the MessageHandlerBase is disposed of, but linking tokens requires memory allocations. Practically speaking, the only caller of these methods is JsonRpc, which itself will cancel the token it explicitly passed to these methods at about the time it disposes of the MessageHandlerBase anyway. So there's no reason to allocate in order to dispose in *both* ways at once.