Add support for ValueTask and ValueTask<T> returning methods
This is *not* more efficient than traditional Task. It's worse. But it allows proxy generation to support interfaces with ValueTask returning methods which is useful because those interfaces MAY not always be remoted. If the interface is there to support remoting but may also work directly, then ValueTask can provide a significant improvement to GC pressure since arbitrary values can be returned without allocating a Task<T> object.

Closes #302