Fixed proxy generation for methods with many parameters
We should not use the short form jump instruction when the instructions between the jump source and target may exceed 127... (IL instructions? bytes?). Anyway, as a rule then, I've switched to long form whenever the jump crosses an IL generating loop, such that we can't be confident that the number of instructions will be small.