Fix possible data corruption from deserializing error details twice
The MessagePackFormatter would deserialize error details twice. The first happened with a live buffer. The second one happened sometime later, concurrently with another thread that is clearing/recycling the buffer. This led to non-deterministic behavior including returning null as the error details but could also lead to reading corrupted data due to use of a recycled buffer.

Fixes #585