Fix race condition in async enum tracker
Calling `BufferBlock<T>.Complete()` can set its own `Completion` property to complete asynchronously. If `ReceiveAsync()` can throw due to completion *before* its `Completion.IsCompleted` is set to `true`, that would repro this problem. I haven't been able to confirm by code inspection or debugger that the `BufferBlock` could do this, but based on the exception you report, I can come to no other conclusion. Based on the documentation of the API, it appears that it's safe to assume than any `InvalidOperationException` is thrown only in the case of a completed block though, so to fix this, I will remove the `when` clause.

Fixes #867